configuration: regexpify rule patterns just once
regexpify() is being called in every request for each of the rules defined.
The rules do not change, so we can just call regexpify() when parsing the
config file to construct a service.
I have profiled APIcast using stapxx and it was clear that this was a problem.
With around 100 rules and 3k rps, regexpify() was consuming like 50% of the CPU
time.