[CVE-2017-7512] reject empty client secret
3scale Service Management API would return 200 when app_key paramater
was empty for legacy purposes. That was not expected by the
gateway which was sending empty key for verification.

This was fixed on 3scale Service Management API and empty app_key
parameter will now return 409.

However, gateway should not rely jsut on 3scale Service Management,
so it performs own validation of the client secret during Access Token
generation.