backend_client: enable no_body option when authorizing
When no_body is enabled via the 3scale-options-header, backend does not
return a response body. The body normally has lots of detail: metrics,
limits, plans, etc. In the normal auth flow, all this information is not
required. We only need to check the status code of the response. By
enabling the no_body option, we will save some work to the 3scale
backend and also reduce network traffic.

no_body cannot be enabled in the oauth flow, as in that case, we need to
parse the response body.