{Dockerfile-development, script/test}: ensure correct lua_modules is used
The problem was that when mounting the volume, ./lua_modules was used
instead of the one generated inside the container.

To solve the issue, lua_modules is moved to a tmp directory when
building the container and we make sure to use it when running the tests
in `script/test`.