policy/local_chain: include export of the local chain in rewrite()
Policies can define an export function to expose the data they need to
share with other policies.

export() worked fine for policies defined in the global chain, but it
did not for policies include in a local (service level) chain. Whatever
was returned in export() did not end up begin in the shared context.
This commit solves the issue.