.circleci/config: delete nightly workflow
Referencing the same job from 2 different workflows is adding suffixes
that break the github integration.
Also, we don't really need this anymore. We added this because we expire
all the images uploaded to quay in a week, but later we added a change
to avoid doing this in images built from the master branch.