t/configuration-loading-boot-remote: delete OIDC test
This is now tested in "configuration-loading-with-oidc.t".
In the tests of this file, we are mocking the "remote_v2" loader, we
cannot properly test loading on boot. That's the reason why the test
deleted in this commit does not add any value over the new ones added in
"configuration-loading-with-oidc.t".