[upstream] fix sending port in the Host header
ref: https://github.com/3scale/apicast/issues/935

The RFC https://tools.ietf.org/html/rfc7230#section-5.4 says:

   A client MUST send a Host header field in all HTTP/1.1 request
   messages.  If the target URI includes an authority component, then a
   client MUST send a field-value for Host that is identical to that
   authority component, excluding any userinfo subcomponent and its "@"
   delimiter (Section 2.7.1).

An "authority component" is defined in in:
* https://tools.ietf.org/html/rfc7230#section-2.7.1
* https://tools.ietf.org/html/rfc3986#section-3.2

That is basically the host [ ":" port ].
This authority section is taken from the normalized URI, that
strips the default port if it matches the scheme. So the port in the
Host header is sent only when it does not match the URI scheme.