Add basic profiling script
The script basically performs the checks I did manually in a previous
PR. Running this script in CircleCI will prevent us from accidentally
adding code that's not jittable.

For now, there's just one basic test. It starts an APIcast in echo-mode
and a second APIcast that just has an Upstream policy that matches any
path and proxies the request to the APicast in echo mode. Then, it
profiles APIcast using the ngx-lj-trace-exits tool.

The idea is to add more tests in the future. Those tests will detect
other kinds of problems with other profiling tools, but also cover more
complex scenarios using several policies, etc.

For now, the code makes a lot of assumptions, but I think that's fine.
We can make it more generic as we add more functionality.